requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
librosa
soundfile
numpy
